requirements.txt
qrcode==7.4.2
pillow==10.2.0
pyinstaller==6.3.0
